// src/lib/huawei/services/geminidb.types.ts
export type GeminiDbInstance = {
  apiType: string;
  actions: string[];
  availabilityZone: string;
  backupKeepDays: string;
  backupPeriod: string;
  backupWindow: string;
  createdAt: string;
  dbUserName: string;
  datastore: string;
  engine: string;
  enterpriseProjectId: string;
  groupCount: number;
  groups: Array<{
    id: string;
    nodes: Array<{
      availabilityZone: string;
      id: string;
      name: string;
      privateIp: string;
      publicIp: string;
      role: string;
      specCode: string;
      status: string;
      subnetId: string;
    }>;
    status: string;
    volumeSize: string;
    volumeUsed: string;
  }>;
  id: string;
  lbIpAddress: string;
  lbPort: string;
  maintenanceWindow: string;
  mode: string;
  name: string;
  nodeCount: number;
  patchAvailable: string;
  payMode: string;
  privateIp: string;
  projectId: string;
  projectName: string;
  publicIp: string;
  region: string;
  securityGroupId: string;
  status: string;
  storage: string;
  subnetId: string;
  timeZone: string;
  updatedAt: string;
  volumeUsed: string;
  vpcId: string;
};